Split MoveCollection controller
